ai-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow>=10.4.0
numpy>=1.26.4
google-generativeai>=0.8.0
reportlab==4.0.7
opencv-python>=4.9.0.80
python-dotenv==1.0.0
pydantic>=2.9.0
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.16.1
scikit-learn>=1.4.0
matplotlib>=3.8.0
